// src/client/src/Components/Generic/Box/types.ts
// Generic
export type Position = 'static' | 'fixed' | 'absolute' | 'sticky' | 'relative';
export type Display =
  | 'inline'
  | 'block'
  | 'flex'
  | 'grid'
  | 'inline-block'
  | 'talbe'
  | 'none';

// Color Control
export type Variant = 'primary' | 'secondary';

// Flexbox
export type AlignItems = 'flex-start' | 'flex-end' | 'center';
export type FlexWrap = 'nowrap' | 'wrap' | 'wrap-reverse';
export type FlexDirection = 'row' | 'row-reverse' | 'column' | 'column-reverse';
export type JustifyContent =
  | 'flex-start'
  | 'flex-end'
  | 'center'
  | 'space-between'
  | 'space-around'
  | 'space-evenly';
export type AlignContent =
  | 'flex-start'
  | 'flex-end'
  | 'center'
  | 'space-between'
  | 'space-around'
  | 'space-evenly';